Add MARL based algorithm
Implement a Multi-agent deep reinforcement learning model to determine
each UE should be connected to VLC AP or WIFI AP.
The allocation of VLC AP adapt a heuristic algorithm.

Each UE is an agent.
Input state to each agent is the
- requrie data rate
- distance to nearest VLC AP
- N_UE in that VLC AP coverage.
Action is a 1-bit vector (0: WiFi, 1:LiFi )
The reward encourages to maximize the system throughput(STP).